requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
Flask==1.1.2
Flask-Compress==1.8.0
Flask-SeaSurf==0.2.2
PyYAML==5.3.1
requests==2.25.0
requests-file==1.5.1
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
selenium==3.141.0
selenium-firefox==0.5.27
youtube-dl==2020.12.5
geckodriver-autoinstaller==0.1.0
